requirements.txt
PyQt5>=5.15.0
opencv-python>=4.5.0
pytesseract>=0.3.8
Pillow>=8.0.0
numpy>=1.19.0
pandas>=1.2.0
PyPDF2>=2.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
customtkinter>=5.2.0
pyperclip>=1.8.2
python-dateutil>=2.8.0